double -> float

+++ vms/c++/observed_float.h
@@ -3,14 +3,14 @@
 
 enum types { U_id, mu_id, F_id, P_id, B_id, S_id, T_id, ntypes };
 
-extern std::vector<double> values[ntypes];
+extern std::vector<float> values[ntypes];
 
 template<enum types T>
 struct od {
 
-    double v;
+    float v;
 
-    od(double d = .0) : v(d) { values[(int)T].push_back(d); }
+    od(float d = .0) : v(d) { values[(int)T].push_back(d); }
 
     template<enum types S>
     od<T> &operator=(od<S> other)
@@ -20,7 +20,7 @@ struct od {
         return *this; 
     }
 
-    explicit operator double() const { return v; }
+    explicit operator float() const { return v; }
 
     template<enum types S>
     od operator-(const od<S> &other) const { return v - other.v; }
